requirements.txt
Django==2.0.2
PyYAML==3.12
Pygments==2.2.0
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
diff-match-patch==20121119
django-cors-headers==2.1.0
django-import-export==1.0.0
djangorestframework==3.7.7
et-xmlfile==1.0.1
httpie==0.9.9
idna==2.6
jdcal==1.3
odfpy==1.3.6
openpyxl==2.5.0
pip==9.0.1
pytz==2018.3
requests==2.18.4
setuptools==38.5.1
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
wheel==0.30.0
xlrd==1.1.0
xlwt==1.3.0